Refactored Matt's elastic search code by abstracting a service
